+++ src/ballots/SchulzeBallot.sol
@@ -5,7 +5,7 @@ import {IBallot} from "./interface/IBallot.sol";
 
 contract SchulzeBallot is IBallot {
     address public electionContract;
-    uint[][] private preferences;
+    uint256[][] private preferences;
 
     modifier onlyOwner() {
         if (msg.sender != electionContract) revert UnathourizedBallot();
@@ -16,24 +16,24 @@ contract SchulzeBallot is IBallot {
         electionContract = _electionAddress;
     }
 
-    function init(uint totalCandidates) external onlyOwner {
-        preferences = new uint[][](totalCandidates);
-        for (uint i = 0; i < totalCandidates; i++) {
-            preferences[i] = new uint[](totalCandidates);
+    function init(uint256 totalCandidates) external onlyOwner {
+        preferences = new uint256[][](totalCandidates);
+        for (uint256 i = 0; i < totalCandidates; i++) {
+            preferences[i] = new uint256[](totalCandidates);
         }
     }
 
-    function vote(uint[] memory voteArr) external onlyOwner {
-        uint totalCandidates = preferences.length;
+    function vote(uint256[] memory voteArr) external onlyOwner {
+        uint256 totalCandidates = preferences.length;
         if (voteArr.length != totalCandidates) revert VoteInputLength();
-        for (uint i = 0; i < totalCandidates; i++) {
-            for (uint j = i + 1; j < totalCandidates; j++) {
+        for (uint256 i = 0; i < totalCandidates; i++) {
+            for (uint256 j = i + 1; j < totalCandidates; j++) {
                 preferences[voteArr[i]][voteArr[j]] += 1;
             }
         }
     }
 
-    function getVotes() external view onlyOwner returns (uint[][] memory) {
+    function getVotes() external view onlyOwner returns (uint256[][] memory) {
         return preferences;
     }
 }
