+++ src/resultCalculators/abstract/VoteWinnerCount.sol
@@ -2,15 +2,13 @@
 pragma solidity ^0.8.20;
 
 abstract contract VoteWinnerCount {
-    function getVoteWinnerCount(
-        uint[] memory candidateList
-    ) internal pure returns (uint, uint) {
-        uint maxVotes = 0;
-        uint winnerCount = 0;
-        uint candidatesLength = candidateList.length;
+    function getVoteWinnerCount(uint256[] memory candidateList) internal pure returns (uint256, uint256) {
+        uint256 maxVotes = 0;
+        uint256 winnerCount = 0;
+        uint256 candidatesLength = candidateList.length;
         // First pass: find the maximum number of votes and the count of candidates with that maximum
-        for (uint i = 0; i < candidatesLength; i++) {
-            uint votes = candidateList[i];
+        for (uint256 i = 0; i < candidatesLength; i++) {
+            uint256 votes = candidateList[i];
             if (votes > maxVotes) {
                 maxVotes = votes;
                 winnerCount = 1;
