+++ src/resultCalculators/interface/IResultCalculator.sol
@@ -1,9 +1,8 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.20;
+
 import {Errors} from "./Errors.sol";
+
 interface IResultCalculator is Errors {
-    function getResults(
-        bytes calldata returnData,
-        uint _resultType
-    ) external returns (uint[] memory);
+    function getResults(bytes calldata returnData, uint256 _resultType) external returns (uint256[] memory);
 }
